+++ packages/apollo-client/src/ApolloClient.ts
@@ -1,7 +1,17 @@
-import { ApolloLink, FetchResult, GraphQLRequest, execute } from 'apollo-link';
+import {
+  ApolloLink,
+  Operation,
+  NextLink,
+  FetchResult,
+  GraphQLRequest,
+  execute,
+} from 'apollo-link';
 import { ExecutionResult } from 'graphql';
 import { ApolloCache, DataProxy } from 'apollo-cache';
-import { isProduction } from 'apollo-utilities';
+import {
+  isProduction,
+  removeConnectionDirectiveFromDocument,
+} from 'apollo-utilities';
 
 import { QueryManager } from './core/QueryManager';
 import { ApolloQueryResult } from './core/types';
@@ -39,6 +49,13 @@ export type ApolloClientOptions<TCacheShape> = {
   defaultOptions?: DefaultOptions;
 };
 
+const supportedDirectives = new ApolloLink(
+  (operation: Operation, forward: NextLink) => {
+    operation.query = removeConnectionDirectiveFromDocument(operation.query);
+    return forward(operation);
+  },
+);
+
 /**
  * This is the primary Apollo Client class. It is used to send GraphQL documents (i.e. queries
  * and mutations) to a GraphQL spec-compliant server over a {@link NetworkInterface} instance,
@@ -64,20 +81,16 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
    *
    * @param link The {@link ApolloLink} over which GraphQL documents will be resolved into a response.
    *
-   * @param initialState The initial state assigned to the store.
-   *
-   * @param initialCache The initial cache to use in the data store.
+   * @param cache The initial cache to use in the data store.
    *
    * @param ssrMode Determines whether this is being run in Server Side Rendering (SSR) mode.
    *
    * @param ssrForceFetchDelay Determines the time interval before we force fetch queries for a
    * server side render.
    *
-   *
    * @param queryDeduplication If set to false, a query will still be sent to the server even if a query
    * with identical parameters (query, variables, operationName) is already in flight.
    *
-   * @param fragmentMatcher A function to use for matching fragment conditions in GraphQL documents
    */
 
   constructor(options: ApolloClientOptions<TCacheShape>) {
@@ -94,13 +107,15 @@ export default class ApolloClient<TCacheShape> implements DataProxy {
     if (!link || !cache) {
       throw new Error(`
         In order to initialize Apollo Client, you must specify link & cache properties on the config object.
+        This is part of the required upgrade when migrating from Apollo Client 1.0 to Apollo Client 2.0.
         For more information, please visit:
           https://apollographql.com/docs/react/setup
         to help you get started.
       `);
     }
 
-    this.link = link;
+    // remove apollo-client supported directives
+    this.link = supportedDirectives.concat(link);
     this.cache = cache;
     this.store = new DataStore(cache);
     this.disableNetworkFetches = ssrMode || ssrForceFetchDelay > 0;